requirements.txt
requests==2.24.0
requests_kerberos==0.14.0
python-dotenv==0.20.0
